Prompt user for Private Passphrase on Wallet Open.
After the user selects the wallet, the app will start
the wallet process, check the selected wallet exists,
and then attempts to open the wallet. To keep the private
passphrase from used in redux, use the state manager
to show the view and pass the entered value to the
rpc call.

Since the PUBLIC PASSPHRASE check already exists, the
PRIVATE PASSPHRASE request will mimic the same setup.

- Split store property OPENWALLET_INPUT into PRIVATE/PUBLIC.
- Replace isInputRequest with differentiator between priv/pub.
- Add private passphrase view to OpenWallet/DecryptForm.